feat: AnimController 분리
