+++ Client/Assets/Scripts/Contents/AnimController.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: c1af6f7cf172469bad421cfcd547b286
+timeCreated: 1710921614
