+++ Client/Assets/Scripts/Contents/AnimController/AlienAnimController.cs
@@ -1,0 +1,34 @@
+﻿using UnityEngine;
+using Fusion;
+
+public class AlienAnimController : BaseAnimController
+{
+    protected override void Init()
+    {
+        base.Init();
+    }
+
+    #region Update
+
+    protected override void PlayIdle()
+    {
+
+    }
+
+    protected override void PlayMove()
+    {
+
+    }
+
+    protected override void PlayInteract()
+    {
+        // TODO
+    }
+
+    protected override void PlayDead()
+    {
+        // TODO
+    }
+
+    #endregion
+}
