+++ Client/Assets/Scripts/Contents/AnimController/AlienAnimController.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 94ff61e00a954d2ca211bbcd85136341
+timeCreated: 1710921660
