+++ Client/Assets/Scripts/Contents/AnimController/CrewAnimController.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3c90107803cc4911a76bf5b4b67d2094
+timeCreated: 1710921649
